+++ cmd/commands/app.go
@@ -17,8 +17,8 @@ import (
 	"github.com/argoproj/argocd-autopilot/pkg/store"
 	"github.com/argoproj/argocd-autopilot/pkg/util"
 
+	argocdv1alpha1 "github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
 	"github.com/ghodss/yaml"
-	memfs "github.com/go-git/go-billy/v5/memfs"
 	billyUtils "github.com/go-git/go-billy/v5/util"
 	"github.com/spf13/cobra"
 	kusttypes "sigs.k8s.io/kustomize/api/types"
@@ -27,19 +27,20 @@ import (
 var (
 	ErrAppAlreadyInstalledOnProject = errors.New("application already installed on project")
 	ErrAppCollisionWithExistingBase = errors.New("an application with the same name and a different base already exists, consider choosing a different name")
+
+	removeAll = billyUtils.RemoveAll
 )
 
 type (
 	AppCreateOptions struct {
-		ProjectName  string
-		FS           fs.FS
-		AppOpts      *application.CreateOptions
-		CloneOptions *git.CloneOptions
+		BaseOptions
+		AppOpts *application.CreateOptions
 	}
-	AppListOptions struct {
-		ProjectName  string
-		FS           fs.FS
-		CloneOptions *git.CloneOptions
+
+	AppDeleteOptions struct {
+		BaseOptions
+		AppName string
+		Global  bool
 	}
 )
 
@@ -50,21 +51,22 @@ func NewAppCommand() *cobra.Command {
 		Short:   "Manage applications",
 		Run: func(cmd *cobra.Command, args []string) {
 			cmd.HelpFunc()(cmd, args)
-			os.Exit(1)
+			exit(1)
 		},
 	}
+	opts, err := addFlags(cmd)
+	die(err)
 
-	cmd.AddCommand(NewAppCreateCommand())
-	cmd.AddCommand(NewAppListCommand())
+	cmd.AddCommand(NewAppCreateCommand(opts))
+	cmd.AddCommand(NewAppListCommand(opts))
+	cmd.AddCommand(NewAppDeleteCommand(opts))
 
 	return cmd
 }
 
-func NewAppCreateCommand() *cobra.Command {
+func NewAppCreateCommand(opts *BaseOptions) *cobra.Command {
 	var (
-		projectName string
-		appOpts     *application.CreateOptions
-		cloneOpts   *git.CloneOptions
+		appOpts *application.CreateOptions
 	)
 
 	cmd := &cobra.Command{
@@ -93,61 +95,41 @@ func NewAppCreateCommand() *cobra.Command {
 			appOpts.AppName = args[0]
 
 			return RunAppCreate(cmd.Context(), &AppCreateOptions{
-				ProjectName:  projectName,
-				FS:           fs.Create(memfs.New()),
-				AppOpts:      appOpts,
-				CloneOptions: cloneOpts,
+				BaseOptions: *opts,
+				AppOpts:     appOpts,
 			})
 		},
 	}
 	appOpts = application.AddFlags(cmd)
-	cloneOpts, err := git.AddFlags(cmd)
-	die(err)
 
-	cmd.Flags().StringVarP(&projectName, "project", "p", "", "Project name")
-
-	die(cmd.MarkFlagRequired("project"))
 	die(cmd.MarkFlagRequired("app"))
 
 	return cmd
 }
 
 func RunAppCreate(ctx context.Context, opts *AppCreateOptions) error {
-	var (
-		err error
-		r   git.Repository
-	)
-
-	log.G().WithFields(log.Fields{
-		"repoURL":  opts.CloneOptions.URL,
-		"revision": opts.CloneOptions.Revision,
-		"appName":  opts.AppOpts.AppName,
-	}).Debug("starting with options: ")
-
-	// clone repo
-	log.G().Infof("cloning git repository: %s", opts.CloneOptions.URL)
-	r, opts.FS, err = opts.CloneOptions.Clone(ctx, opts.FS)
+	r, repofs, err := prepareRepo(ctx, &opts.BaseOptions)
 	if err != nil {
 		return err
 	}
-	log.G().Infof("using revision: \"%s\", installation path: \"%s\"", opts.CloneOptions.Revision, opts.FS.Root())
 
-	if !opts.FS.ExistsOrDie(store.Default.BootsrtrapDir) {
-		return fmt.Errorf(util.Doc("Bootstrap folder not found, please execute `<BIN> repo bootstrap --installation-path %s` command"), opts.FS.Root())
+	if opts.AppOpts.DestServer == store.Default.DestServer {
+		opts.AppOpts.DestServer, err = getProjectDestServer(repofs, opts.ProjectName)
+		if err != nil {
+			return err
+		}
 	}
 
-	projectExists := opts.FS.ExistsOrDie(opts.FS.Join(store.Default.ProjectsDir, opts.ProjectName+".yaml"))
-	if !projectExists {
-		return fmt.Errorf(util.Doc("project '%[1]s' not found, please execute `<BIN> project create %[1]s`"), opts.ProjectName)
+	if opts.AppOpts.DestNamespace == "" {
+		opts.AppOpts.DestNamespace = "default"
 	}
-	log.G().Debug("repository is ok")
 
 	app, err := opts.AppOpts.Parse()
 	if err != nil {
 		return fmt.Errorf("failed to parse application from flags: %v", err)
 	}
 
-	if err = createApplicationFiles(opts.FS, app, opts.ProjectName); err != nil {
+	if err = createApplicationFiles(repofs, app, opts.ProjectName); err != nil {
 		if errors.Is(err, ErrAppAlreadyInstalledOnProject) {
 			return fmt.Errorf("application '%s' already exists in project '%s': %w", app.Name(), opts.ProjectName, ErrAppAlreadyInstalledOnProject)
 		}
@@ -156,9 +138,10 @@ func RunAppCreate(ctx context.Context, opts *AppCreateOptions) error {
 	}
 
 	log.G().Info("committing changes to gitops repo...")
-	if err = r.Persist(ctx, &git.PushOptions{CommitMsg: getCommitMsg(opts)}); err != nil {
+	if err = r.Persist(ctx, &git.PushOptions{CommitMsg: getCommitMsg(opts, repofs)}); err != nil {
 		return fmt.Errorf("failed to push to repo: %w", err)
 	}
+
 	log.G().Infof("installed application: %s", opts.AppOpts.AppName)
 
 	return nil
@@ -193,6 +176,7 @@ func createApplicationFiles(repoFS fs.FS, app application.Application, projectNa
 	if err != nil {
 		return fmt.Errorf("failed to marshal app overlay kustomization: %w", err)
 	}
+
 	if exists, err := writeApplicationFile(repoFS, overlayKustomizationPath, "overlay", overlayKustomizationYAML); err != nil {
 		return err
 	} else if exists {
@@ -225,6 +209,7 @@ func createApplicationFiles(repoFS fs.FS, app application.Application, projectNa
 	if err != nil {
 		return fmt.Errorf("failed to marshal app config.json: %w", err)
 	}
+
 	if _, err = writeApplicationFile(repoFS, configPath, "config", config); err != nil {
 		return err
 	}
@@ -260,24 +245,40 @@ func writeApplicationFile(repoFS fs.FS, path, name string, data []byte) (bool, e
 		log.G().Infof("'%s' file exists in '%s'", name, absPath)
 		return true, nil
 	}
+
 	log.G().Infof("created '%s' file at '%s'", name, absPath)
 	return false, nil
 }
 
-func getCommitMsg(opts *AppCreateOptions) string {
+func getCommitMsg(opts *AppCreateOptions, repofs fs.FS) string {
 	commitMsg := fmt.Sprintf("installed app '%s' on project '%s'", opts.AppOpts.AppName, opts.ProjectName)
-	if opts.CloneOptions.RepoRoot != "" {
-		commitMsg += fmt.Sprintf(" installation-path: '%s'", opts.CloneOptions.RepoRoot)
+	if repofs.Root() != "" {
+		commitMsg += fmt.Sprintf(" installation-path: '%s'", repofs.Root())
 	}
+
 	return commitMsg
 }
-func NewAppListCommand() *cobra.Command {
-	var (
-		projectName string
-		//	appListOpts   *application.
-		cloneOpts *git.CloneOptions
-	)
 
+var getProjectDestServer = func(repofs fs.FS, projectName string) (string, error) {
+	f, err := repofs.Open(repofs.Join(store.Default.ProjectsDir, projectName+".yaml"))
+	if err != nil {
+		return "", err
+	}
+
+	d, err := ioutil.ReadAll(f)
+	if err != nil {
+		return "", fmt.Errorf("failed to read namespace file: %w", err)
+	}
+
+	p := &argocdv1alpha1.AppProject{}
+	if err = yaml.Unmarshal(d, p); err != nil {
+		return "", fmt.Errorf("failed to unmarshal project: %w", err)
+	}
+
+	return p.Annotations[store.Default.DestServerAnnotation], nil
+}
+
+func NewAppListCommand(opts *BaseOptions) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "list [PROJECT_NAME]",
 		Short: "List all applications in a project",
@@ -300,89 +301,159 @@ func NewAppListCommand() *cobra.Command {
 			if len(args) < 1 {
 				log.G().Fatal("must enter a project name")
 			}
-			projectName = args[0]
+			opts.ProjectName = args[0]
 
-			return RunAppList(cmd.Context(), &AppListOptions{
-				ProjectName:  projectName,
-				FS:           fs.Create(memfs.New()),
-				CloneOptions: cloneOpts,
-			})
+			return RunAppList(cmd.Context(), opts)
 		},
 	}
-	cloneOpts, err := git.AddFlags(cmd)
-	util.Die(err)
 
 	return cmd
 }
 
-func RunAppList(ctx context.Context, opts *AppListOptions) error {
-
-	var (
-		err error
-	)
-
-	log.G().WithFields(log.Fields{
-		"repoURL":  opts.CloneOptions.URL,
-		"revision": opts.CloneOptions.Revision,
-	}).Debug("starting with options: ")
-
-	// clone repo
-	log.G().Infof("cloning git repository: %s", opts.CloneOptions.URL)
-	_, opts.FS, err = opts.CloneOptions.Clone(ctx, opts.FS)
+func RunAppList(ctx context.Context, opts *BaseOptions) error {
+	_, repofs, err := prepareRepo(ctx, opts)
 	if err != nil {
 		return err
 	}
 
-	log.G().Infof("using revision: \"%s\", installation path: \"%s\"", opts.CloneOptions.Revision, opts.FS.Root())
-	if !opts.FS.ExistsOrDie(store.Default.BootsrtrapDir) {
-		log.G().Fatalf("Bootstrap folder not found, please execute `repo bootstrap --installation-path %s` command", opts.FS.Root())
-	}
-
-	projExists := opts.FS.ExistsOrDie(opts.FS.Join(store.Default.ProjectsDir, opts.ProjectName+".yaml"))
-	if !projExists {
-		log.G().Fatalf(util.Doc(fmt.Sprintf("project '%[1]s' not found, please execute `<BIN> project create %[1]s`", opts.ProjectName)))
-	}
-
-	log.G().Debug("repository is ok")
-
 	// get all apps beneath kustomize <project>\overlayes
-
-	matches, err := billyUtils.Glob(opts.FS, fmt.Sprintf("/kustomize/*/overlays/%s", opts.ProjectName))
+	glob := repofs.Join(store.Default.KustomizeDir, "*", store.Default.OverlaysDir, opts.ProjectName)
+	matches, err := billyUtils.Glob(repofs, glob)
 	if err != nil {
 		log.G().Fatalf("failed to run glob on %s", opts.ProjectName)
 	}
+
 	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', 0)
 	_, _ = fmt.Fprintf(w, "PROJECT\tNAME\tDEST_NAMESPACE\tDEST_SERVER\t\n")
 
 	for _, appName := range matches {
-
-		conf, err := getConfigFileFromPath(opts.FS, appName)
+		conf, err := getConfigFileFromPath(repofs, appName)
 		if err != nil {
 			return err
 		}
 
 		fmt.Fprintf(w, "%s\t%s\t%s\t%s\t\n", opts.ProjectName, conf.UserGivenName, conf.DestNamespace, conf.DestServer)
 	}
+
 	_ = w.Flush()
 	return nil
-
 }
-func getConfigFileFromPath(fs fs.FS, appName string) (*application.Config, error) {
 
+func getConfigFileFromPath(fs fs.FS, appName string) (*application.Config, error) {
 	confFileName := fmt.Sprintf("%s/config.json", appName)
 	file, err := fs.Open(confFileName)
 	if err != nil {
 		return nil, fmt.Errorf("%s not found", confFileName)
 	}
+
 	b, err := ioutil.ReadAll(file)
 	if err != nil {
 		return nil, fmt.Errorf("failed to read file %s", confFileName)
 	}
+
 	conf := application.Config{}
 	err = json.Unmarshal(b, &conf)
 	if err != nil {
 		return nil, fmt.Errorf("failed to unmarshal file %s", confFileName)
 	}
-	return &conf, nil
 
+	return &conf, nil
+}
+
+func NewAppDeleteCommand(opts *BaseOptions) *cobra.Command {
+	var (
+		appName string
+		global  bool
+	)
+
+	cmd := &cobra.Command{
+		Use:   "delete [APP_NAME]",
+		Short: "Delete an application from a project",
+		Example: util.Doc(`
+# To run this command you need to create a personal access token for your git provider,
+# and have a bootstrapped GitOps repository, and provide them using:
+	
+		export GIT_TOKEN=<token>
+		export GIT_REPO=<repo_url>
+
+# or with the flags:
+	
+		--token <token> --repo <repo_url>
+		
+# Get list of installed applications in a specifc project
+	
+	<BIN> app delete <app_name> --project <project_name>
+`),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) < 1 {
+				log.G().Fatal("must enter application name")
+			}
+
+			appName = args[0]
+
+			if opts.ProjectName == "" && !global {
+				log.G().Fatal("must enter project name OR use '--global' flag")
+			}
+
+			return RunAppDelete(cmd.Context(), &AppDeleteOptions{
+				BaseOptions: *opts,
+				AppName:     appName,
+				Global:      global,
+			})
+		},
+	}
+
+	cmd.Flags().BoolVarP(&global, "global", "g", false, "global")
+
+	return cmd
+}
+
+func RunAppDelete(ctx context.Context, opts *AppDeleteOptions) error {
+	r, repofs, err := prepareRepo(ctx, &opts.BaseOptions)
+	if err != nil {
+		return err
+	}
+
+	appDir := repofs.Join(store.Default.KustomizeDir, opts.AppName)
+	appExists := repofs.ExistsOrDie(appDir)
+	if !appExists {
+		return fmt.Errorf(util.Doc(fmt.Sprintf("application '%s' not found", opts.AppName)))
+	}
+
+	var dirToRemove string
+	commitMsg := fmt.Sprintf("Deleted app '%s'", opts.AppName)
+	if opts.Global {
+		dirToRemove = appDir
+	} else {
+		appOverlaysDir := repofs.Join(appDir, store.Default.OverlaysDir)
+		projectDir := repofs.Join(appOverlaysDir, opts.ProjectName)
+		overlayExists := repofs.ExistsOrDie(projectDir)
+		if !overlayExists {
+			return fmt.Errorf(util.Doc(fmt.Sprintf("application '%s' not found in project '%s'", opts.AppName, opts.ProjectName)))
+		}
+
+		allOverlays, err := repofs.ReadDir(appOverlaysDir)
+		if err != nil {
+			return fmt.Errorf("failed to read overlays directory '%s': %w", appOverlaysDir, err)
+		}
+
+		if len(allOverlays) == 1 {
+			dirToRemove = appDir
+		} else {
+			commitMsg += fmt.Sprintf(" from project '%s'", opts.ProjectName)
+			dirToRemove = projectDir
+		}
+	}
+
+	err = removeAll(repofs, dirToRemove)
+	if err != nil {
+		return fmt.Errorf("failed to delete directory '%s': %w", dirToRemove, err)
+	}
+
+	log.G().Info("committing changes to gitops repo...")
+	if err = r.Persist(ctx, &git.PushOptions{CommitMsg: commitMsg}); err != nil {
+		return fmt.Errorf("failed to push to repo: %w", err)
+	}
+
+	return nil
 }