+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog:
 
+# v0.1.5
+* doc fixes
+* you no longer need to run `argocd login` after running `repo bootstrap`
+* added `app delete` command
+
+# v0.1.4
+* doc fixes
+* fixed adding application on a remote cluster
+
 # v0.1.3
 * fixed docker image creation