+++ docs/releases/release_notes.md
@@ -1,6 +1,7 @@
 ## Changes:
-### Images:
-* fixed image creation
+* doc fixes
+* you no longer need to run `argocd login` after running `repo bootstrap`
+* added `app delete` command
 
 ## Installation: